fix: gRPC message size type, empty command validation, temp file leak, DbError variants
- snip-sync: Change grpc_max_message_size from usize to u32 throughout
  (config, env var parsing, tonic call) to prevent silent truncation on
  values >4 GiB. Tonic's max_frame_size already expects u32.
- snip-sync: Add empty command validation in validate_snippet — reject
  snippets with empty or whitespace-only commands, matching client-side
  Snippet::new() behavior and premade.rs filtering precedent.
- snip-sync: Extract hardcoded rate limiter window (60s) into WINDOW_SECS
  constant. Used in cleanup task, load_state, and save_state pruning
  for explicit relationship between values.
- snip-sync: Add DbError::Validation and DbError::Internal variants.
  Validation replaces Conflict for invalid library names (maps to
  INVALID_ARGUMENT instead of ALREADY_EXISTS). Internal replaces
  Conflict for crypto failures in hash_api_key (maps to INTERNAL).
- src: Add TempFileGuard RAII struct that deletes temp file on drop
  unless persist() is called. Applied to all 5 atomic-write save
  functions (save_library, save_config, save_sync_settings,
  save_snippets, ThemeManager::save_config) to prevent orphaned temp
  files when write fails.

+++ snip-sync/src/main.rs
@@ -268,7 +268,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let grpc_handle = tokio::spawn(async move {
         let server = tonic::transport::Server::builder()
             .timeout(timeout)
-            .max_frame_size(grpc_max_message_size as u32)
+            .max_frame_size(grpc_max_message_size)
             .add_service(SnippetSyncServer::new(grpc_service))
             .serve(grpc_addr);
